Adding the Log manager with DB schema to monitor connections
When the DB is deployed and the Log Manager hosted on some server (now using Flask) all new connections are saved in the DB by ssh_module.py. The log_manager.py then fetches them and displays on the web dashboard that also needs to be hosted somewhere. Web dashboard is a simple web app that displays the tables of the connections to shelLM honeypot, as well the history of the session in the Commands_and_Answers table.